requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
python-docx==0.8.11
sentence-transformers==2.2.2
torch==2.0.1
numpy==1.24.3
pandas==2.0.3
faiss-cpu==1.7.4
openai==1.3.3
python-dotenv==1.0.0
nltk==3.8.1
Werkzeug==2.3.7
gunicorn==21.2.0
